+++ ep-sdk/src/services/EpSdkApplicationDomainsService.ts
@@ -1,4 +1,9 @@
-import { EpSdkApiContentError, EpSdkServiceError } from '../EpSdkErrors';
+/**
+ * This is the doc comment for EpSdkApplicationDomainsService.ts
+ *
+ * @module EpSdkApplicationDomainsService
+ */
+ import { EpSdkApiContentError, EpSdkServiceError } from '../EpSdkErrors';
 import { EpSdkLogger } from '../EpSdkLogger';
 import { EEpSdkLoggerCodes } from '../EpSdkLoggerCodes';
 import { 
@@ -6,10 +11,10 @@ import {
   ApplicationDomainResponse,
   ApplicationDomainsResponse,
   ApplicationDomainsService,
-} from '@solace-iot-team/ep-openapi-node';
+} from '@solace-labs/ep-openapi-node';
 import { EpSdkService } from './EpSdkService';
 
-class EpSdkApplicationDomainsService extends EpSdkService {
+export class EpSdkApplicationDomainsService extends EpSdkService {
 
   public getByName = async({ applicationDomainName }:{
     applicationDomainName: string;
@@ -84,6 +89,5 @@ class EpSdkApplicationDomainsService extends EpSdkService {
   }
 
 }
-
 export default new EpSdkApplicationDomainsService();