+++ ep-sdk/src/services/EpSdkEnumVersionsService.ts
@@ -6,10 +6,14 @@ import {
   EnumVersionResponse,
   EnumVersionsResponse,
   VersionedObjectStateChangeRequest
-} from '@solace-iot-team/ep-openapi-node';
+} from '@solace-labs/ep-openapi-node';
 import EpSdkEnumsService from "./EpSdkEnumsService";
 import { EpSdkVersionService } from "./EpSdkVersionService";
 
+/**
+ * @category Services
+ * @group Services
+ */
 class EpSdkEnumVersionsService extends EpSdkVersionService {
 
   public getVersionByVersion = async ({ enumId, enumVersionString }: {