+++ ep-sdk/test/specs/services/applicationDomainsService.spec.ts
@@ -5,7 +5,7 @@ import { TestLogger } from '../../lib/TestLogger';
 import { TestContext } from '../../lib/TestContext';
 import TestConfig from '../../lib/TestConfig';
 import { TestUtils } from '../../lib/TestUtils';
-import { ApiError, ApplicationDomain, ApplicationDomainResponse, ApplicationDomainsService } from '@solace-iot-team/ep-openapi-node';
+import { ApiError, ApplicationDomain, ApplicationDomainResponse, ApplicationDomainsService } from '@solace-labs/ep-openapi-node';
 import EpSdkApplicationDomainsService from '../../../src/services/EpSdkApplicationDomainsService';
 import { EpSdkError, EpSdkServiceError } from '../../../src/EpSdkErrors';